+++ tests/examples/test_delete_rows_ignore_unspecified/migrations/versions/a63ad8efc2f4_.py
@@ -5,6 +5,7 @@ Revises:
 Create Date: 2022-11-07 11:31:40.736066
 
 """
+
 import sqlalchemy as sa
 from alembic import op