+++ tests/examples/test_single_row_ops/migrations/versions/34137c7892d1_.py
@@ -5,6 +5,7 @@ Revises: a63ad8efc2f4
 Create Date: 2023-01-04 16:07:48.335881
 
 """
+
 from alembic import op
 
 # revision identifiers, used by Alembic.